Created files for all methods and functions from index.js
