+++ test/e2e/support/utils/accept-page-dialog.js
@@ -1,0 +1,8 @@
+/**
+ * Callback which automatically accepts dialog.
+ *
+ * @param {puppeteer.Dialog} dialog Dialog object dispatched by page via the 'dialog' event.
+ */
+export async function acceptPageDialog( dialog ) {
+	await dialog.accept();
+}
