+++ test/e2e/support/utils/are-pre-publish-checks-enabled.js
@@ -1,0 +1,5 @@
+export async function arePrePublishChecksEnabled() {
+	return page.evaluate( () =>
+		window.wp.data.select( 'core/editor' ).isPublishSidebarEnabled()
+	);
+}
