+++ test/e2e/support/utils/clear-local-storage.js
@@ -1,0 +1,3 @@
+export async function clearLocalStorage() {
+	await page.evaluate( () => window.localStorage.clear() );
+}
