+++ test/e2e/support/utils/click-block-appender.js
@@ -1,0 +1,6 @@
+/**
+ * Clicks the default block appender.
+ */
+export async function clickBlockAppender() {
+	await page.click( '.editor-default-block-appender__content' );
+}
