+++ test/e2e/support/utils/click-on-close-modal-button.js
@@ -1,0 +1,19 @@
+/**
+ * Click on the close button of an open modal.
+ *
+ * @param {?string} modalClassName Class name for the modal to close
+ */
+export async function clickOnCloseModalButton( modalClassName ) {
+	let closeButtonClassName =
+		'.components-modal__header .components-icon-button';
+
+	if ( modalClassName ) {
+		closeButtonClassName = `${ modalClassName } ${ closeButtonClassName }`;
+	}
+
+	const closeButton = await page.$( closeButtonClassName );
+
+	if ( closeButton ) {
+		await page.click( closeButtonClassName );
+	}
+}
