+++ test/e2e/support/utils/convert-block.js
@@ -1,0 +1,6 @@
+export async function convertBlock( name ) {
+	await page.mouse.move( 200, 300, { steps: 10 } );
+	await page.mouse.move( 250, 350, { steps: 10 } );
+	await page.click( '.editor-block-switcher__toggle' );
+	await page.click( `.editor-block-types-list__item[aria-label="${ name }"]` );
+}
