+++ test/e2e/support/utils/disable-pre-publish-checks.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import { toggleOption } from './toggle-option';
+
+export async function disablePrePublishChecks() {
+	await toggleOption( 'Enable Pre-publish Checks', false );
+}
