+++ test/e2e/support/utils/enable-pre-publish-checks.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import { toggleOption } from './toggle-option';
+
+export async function enablePrePublishChecks() {
+	await toggleOption( 'Enable Pre-publish Checks', true );
+}
