+++ test/e2e/support/utils/go-to-wp-path.js
@@ -1,0 +1,8 @@
+/**
+ * Internal dependencies
+ */
+import { getUrl } from './get-url';
+
+export async function goToWPPath( WPPath, query ) {
+	await page.goto( getUrl( WPPath, query ) );
+}
