+++ test/e2e/support/utils/is-wp-path.js
@@ -1,0 +1,12 @@
+/**
+ * Internal dependencies
+ */
+import { getUrl } from './get-url';
+
+export function isWPPath( WPPath, query = '' ) {
+	const currentUrl = new URL( page.url() );
+
+	currentUrl.search = query;
+
+	return getUrl( WPPath ) === currentUrl.href;
+}
