+++ test/e2e/support/utils/meta-key.js
@@ -1,0 +1,8 @@
+/**
+ * Platform-specific meta key.
+ *
+ * @see pressWithModifier
+ *
+ * @type {string}
+ */
+export const META_KEY = process.platform === 'darwin' ? 'Meta' : 'Control';
