+++ test/e2e/support/utils/open-publish-panel.js
@@ -1,0 +1,11 @@
+/**
+ * Opens the publish panel.
+ */
+export async function openPublishPanel() {
+	await page.click( '.editor-post-publish-panel__toggle' );
+
+	// Disable reason: Wait for the animation to complete, since otherwise the
+	// click attempt may occur at the wrong point.
+	// eslint-disable-next-line no-restricted-syntax
+	await page.waitFor( 100 );
+}
