+++ test/e2e/support/utils/publish-post.js
@@ -1,0 +1,20 @@
+/**
+ * Internal dependencies
+ */
+import { openPublishPanel } from './open-publish-panel';
+
+/**
+ * Publishes the post, resolving once the request is complete (once a notice
+ * is displayed).
+ *
+ * @return {Promise} Promise resolving when publish is complete.
+ */
+export async function publishPost() {
+	await openPublishPanel();
+
+	// Publish the post
+	await page.click( '.editor-post-publish-button' );
+
+	// A success notice should show up
+	return page.waitForSelector( '.components-notice.is-success' );
+}
