+++ test/e2e/support/utils/select-block-by-client-id.js
@@ -1,0 +1,10 @@
+/**
+ * Given the clientId of a block, selects the block on the editor.
+ *
+ * @param {string} clientId Identified of the block.
+ */
+export async function selectBlockByClientId( clientId ) {
+	await page.evaluate( ( id ) => {
+		wp.data.dispatch( 'core/editor' ).selectBlock( id );
+	}, clientId );
+}
