+++ test/e2e/support/utils/switch-to-test-user.js
@@ -1,0 +1,25 @@
+/**
+ * Internal dependencies
+ */
+import { goToWPPath } from './go-to-wp-path';
+import { login } from './login';
+
+const WP_ADMIN_USER = {
+	username: 'admin',
+};
+
+const {
+	WP_USERNAME = WP_ADMIN_USER.username,
+} = process.env;
+
+/**
+ * Switches the current user to whichever user we should be
+ * running the tests as (if we're not already that user).
+ */
+export async function switchToTestUser() {
+	if ( WP_USERNAME === WP_ADMIN_USER.username ) {
+		return;
+	}
+	await goToWPPath( 'wp-login.php' );
+	await login();
+}
